requirements.txt
appnope==0.1.4
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
huggingface-hub==0.21.2
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
prompt-toolkit==3.0.42
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tokenizers==0.15.2
torch==1.13.1
torch-summary==1.4.5
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.41.2
widgetsnbextension==4.0.10
zipp==3.17.0
